feat: add /login route
