--- src/logging/error-logs.log
@@ -1,2 +1,0 @@
-2021-06-27 10:00:05 Response status code: 400 
-Response message: Route not found
