+++ src/resources/login/login.router.ts
@@ -3,6 +3,6 @@ import * as controller from './login.controller';
 
 const router = Router();
 
-router.route('/login').post(controller.login);
+router.route('/').post(controller.login);
 
 export { router };
