+++ src/resources/login/login.service.ts
@@ -1,0 +1,6 @@
+import * as loginRepo from './login.memory.repository';
+import { User } from '../users/user.model';
+
+const authUser = async (user: User): Promise<User> => loginRepo.authUser(user); 
+
+export { authUser };
