+++ src/resources/users/user.router.ts
@@ -4,7 +4,6 @@ import { hashPass } from '../../middleware/hashPass';
 
 const router = Router();
 
-router.route('/login').post(controller.login);
 router.route('/').get(controller.getAll);
 router.route('/:id').get(controller.getById);
 router.route('/').post(hashPass, controller.createUser);
